+++ quoridor/src/main.rs
@@ -1,12 +1,6 @@
 
-use actix_web::{middleware, web, App, HttpRequest, HttpServer};
-use actix_web::body::{Body, ResponseBody};
-
-
-pub async fn index(req: HttpRequest) -> &'static str {
-    println!("REQ: {:?}", req);
-    "Hello world!"
-}
+use actix_web::{middleware, web, App, HttpServer};
+mod handler;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
@@ -17,63 +11,45 @@ async fn main() -> std::io::Result<()> {
         App::new()
             // enable logger
             .wrap(middleware::Logger::default())
-            .service(web::resource("/").to(index))
+            .service(web::resource("/").to(handler::hello))
     })
     .bind("api:8383")?
     .run()
     .await
 }
 
-trait BodyTest {
-    fn as_str(&self) -> &str;
-}
-
-impl BodyTest for ResponseBody<Body> {
-    fn as_str(&self) -> &str {
-        match self {
-            ResponseBody::Body(ref b) => match b {
-                Body::Bytes(ref by) => std::str::from_utf8(&by).unwrap(),
-                _ => panic!(),
-            },
-            ResponseBody::Other(ref b) => match b {
-                Body::Bytes(ref by) => std::str::from_utf8(&by).unwrap(),
-                _ => panic!(),
-            },
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use actix_web::dev::Service;
-    use actix_web::{http, test, web, App, Error};
+    use actix_web::{test, web, App};
+    use actix_web::body::{Body, ResponseBody};
 
-    #[actix_rt::test]
-    async fn test_hello_handler() -> Result<(), Error> {
-        let app = App::new().route("/", web::get().to(index));
-        let mut app = test::init_service(app).await;
-
-        let req = test::TestRequest::get().uri("/").to_request();
-        let resp = app.call(req).await.unwrap();
-
-        assert_eq!(resp.status(), http::StatusCode::OK);
-
-        let response_body = match resp.response().body().as_ref() {
-            Some(actix_web::body::Body::Bytes(bytes)) => bytes,
-            _ => panic!("Response error"),
-        };
-
-        assert_eq!(response_body, r##"Hello world!"##);
-
-        Ok(())
+    trait BodyTest {
+        fn as_str(&self) -> &str;
+    }
+    
+    impl BodyTest for ResponseBody<Body> {
+        fn as_str(&self) -> &str {
+            match self {
+                ResponseBody::Body(ref b) => match b {
+                    Body::Bytes(ref by) => std::str::from_utf8(&by).unwrap(),
+                    _ => panic!(),
+                },
+                ResponseBody::Other(ref b) => match b {
+                    Body::Bytes(ref by) => std::str::from_utf8(&by).unwrap(),
+                    _ => panic!(),
+                },
+            }
+        }
     }
 
     #[actix_rt::test]
     async fn test_get_hello_world() {
-        let mut app = test::init_service(App::new().route("/", web::get().to(index))).await;
+        let mut app = test::init_service(App::new().route("/", web::get().to(handler::hello))).await;
         let req = test::TestRequest::with_header("content-type", "text/plain").to_request();
+        
         let resp = test::call_service(&mut app, req).await;
+        
         assert!(resp.status().is_success());
         assert_eq!(resp.response().body().as_str(), "Hello world!");
     }
